Added few log functions

+++ libui/src/ui_main/ui_main_loop.c
@@ -6,7 +6,7 @@
 /*   By: sbednar <sbednar@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/30 21:26:26 by sbednar           #+#    #+#             */
-/*   Updated: 2019/04/22 06:06:12 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/04/22 10:50:55 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,6 +33,6 @@ void	ui_main_loop(t_ui_main *m)
 		// SDL_PollEvent(&(m->sdl_event));
 		// ui_main_handle_event(m);
 		// ui_draw(m);
-		SDL_Delay(1000 / FPS);
+		SDL_Delay((Uint32)(1000.0f / FPS));
 	}
 }
